backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.2
python-dotenv==1.0.0
paho-mqtt==1.6.1
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
alembic==1.12.1
psycopg2-binary==2.9.9
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2
scikit-learn==1.3.2
pyjwt==2.8.0
httpx==0.25.2
motor==3.3.1
onnxruntime==1.16.0
prophet==1.1.4
xgboost==2.0.1
supabase==2.1.0
modbus-tk==1.1.3
pymodbus==3.5.4
